CustomString/Math: Fixed compilation error with MinGW64
git-svn-id: file:///home/michael/svn/yuni/trunk@1428 7e728022-c0ab-4603-8e65-c666ead28a33

+++ src/yuni/core/io/file/stream.cpp
@@ -32,7 +32,7 @@ namespace File
 		if (!len)
 			return NULL;
 		wchar_t* buffer = new wchar_t[len + 2];
-		const int n = MultiByteToWideChar(CP_UTF8, 0, filename, len, buffer, len);
+		const int n = MultiByteToWideChar(CP_UTF8, 0, filename, static_cast<int>(len), buffer, static_cast<int>(len));
 		if (n <= 0)
 		{
 			delete[] buffer;
